refactor and streamline layouts, update metadata assets
