+++ src/app/(default)/about/content.ts
