+++ src/app/layout.tsx
@@ -2,13 +2,15 @@ import { cn } from "@/lib/utils";
 import { GeistSans, GeistMono } from "geist/font";
 import { Analytics } from "@vercel/analytics/react";
 import ThemeProvider from "@/components/theme-provider";
-import SiteHeader from "@/components/site-header";
-import SiteFooter from "@/components/site-footer";
 import "@/styles/globals.css";
 
 export const metadata = {
-  title: "kalynbeach",
+  title: {
+    template: "%s | kalynbeach",
+    default: "kalynbeach",
+  },
   description: "Kalyn Beach",
+  metadataBase: new URL("https://kalynbeach.xyz"),
 };
 
 export default function RootLayout({
@@ -25,13 +27,7 @@ export default function RootLayout({
           enableSystem
           disableTransitionOnChange
         >
-          <div className="w-full md:max-w-3xl lg:max-w-4xl min-h-screen mx-auto flex flex-col">
-            <SiteHeader />
-            <main className="min-h-max p-8 md:px-4 md:py-12 flex-1 flex flex-col">
-              {children}
-            </main>
-            <SiteFooter />
-          </div>
+          {children}
         </ThemeProvider>
         <Analytics />
       </body>
